Renamed the project and rephrased to position it
There are a ton of podcast downloader scripts on PYPI, there are simple ones and
complicated ones, but neither of them does what this: simply dowloading MP3
files without any hassle. I made this clear in the README and in help, so I want
to "position" it as the simplest podcast downloader you can get.