Refactored site_parser, added more tests.
Now we need to take account the URL to filename parser when parsing site names,
and pass it to the Episode class.

Refactored parse_site, and moved to site_parser.py, added more tests for the
more complex cases and naked domains.

Inform the user with a message which Podcast will be downloaded to avoid
surprises. It is possible to give an invalid URL or domain for a podcast, but
the parser will still find it, but hopefully it won't be a problem. Could handle
those cases more strictly, but this seems to be more flexible and convenient.