Handle CTRL-C better.
This is the best we can do, because ThreadPoolexecutor wait for the unfinished
threads anyway. I tried with executor.shutdown(wait=False) without much success.
This is the most responsie solution I could come up with.

The main point is to submit jobs for the Executor in batches, because otherwise
it would wait for ALL of the submitted jobs to finish.

The grouper recipe came from the itertools Python help page:
https://docs.python.org/3/library/itertools.html#itertools-recipes