Fine-tuned help messages and output, fixed README
Log the finished (downloaded) filename instead of full path.
It might be confusing because it can be a relative path, but at the start, we
write the full absolute path of the download dir. Printing the absolute path
here would be too verbose, filename is sufficient.

Fine-tuned "searching episode" message.

Small help change to clarify directory default.

TEMP: readme