Better default for download_dir (Podcast.name). Fixes #11.
if someone run podcast-dl talkpython.fm the first time and podcast-dl talkpython
for the second time, they want to download the same show, but in this case it
will download into two different folders. Would be nice if it would go into the
same folder by default. Of course the -d option is there, it's just about a
nice default.