Use semaphore instead of grouper
It is constantly keeping max_threads number of downloads vs
deplete a group and start again, so it should be faster than the grouper.
Also the progressbar is way more precise this way.

The grouper function and tests has been deleted.

The README has been rephrased a bit for the --max-threads option,
because of the changed behaviour.